+++ src/app/lists/[listId]/movieCatalogue.tsx
@@ -1,5 +1,4 @@
 import Link from "next/link";
-import { Star } from "react-feather";
 
 type MovieCatalogueItem = {
   id: number;
